moved auth inside apps to replace old routes
