--- frontend/src/Auth/Login/page.tsx
@@ -1,1 +1,0 @@
-// import { useSignIn } from "@clerk/nextjs";
