+++ frontend/src/app/Auth/Login/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { useSignIn } from "@clerk/nextjs";
-import Login from "../../components/Login";
+import Login from "../../../components/Login";
 
 export default function Page() {
   const { isLoaded, signIn, setActive } = useSignIn();
